requirements.txt

django-modeltranslation
